Drop support for PHP 7.0
