+++ src/test/php/PDepend/Bugs/NamespacedConstsAndFunctionsBug00000247Test.php
@@ -117,7 +117,7 @@ class NamespacedConstsAndFunctionsBug00000247Test extends AbstractRegressionTest
     protected function createPHPParser(Tokenizer $tokenizer, Builder $builder, CacheDriver $cache)
     {
         return $this->getAbstractClassMock(
-            'PDepend\\Source\\Language\\PHP\\PHPParserVersion70',
+            'PDepend\\Source\\Language\\PHP\\PHPParserVersion71',
             array($tokenizer, $builder, $cache)
         );
     }
