--- src/test/php/PDepend/Source/Language/PHP/PHPParserVersion70Test.php
@@ -1,761 +1,0 @@
-<?php
-/**
- * This file is part of PDepend.
- *
- * PHP Version 5
- *
- * Copyright (c) 2008-2017 Manuel Pichler <mapi@pdepend.org>.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- *   * Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- *
- *   * Redistributions in binary form must reproduce the above copyright
- *     notice, this list of conditions and the following disclaimer in
- *     the documentation and/or other materials provided with the
- *     distribution.
- *
- *   * Neither the name of Manuel Pichler nor the names of his
- *     contributors may be used to endorse or promote products derived
- *     from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- * @copyright 2008-2017 Manuel Pichler. All rights reserved.
- * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- * @since 2.3
- */
-
-namespace PDepend\Source\Language\PHP;
-
-use PDepend\AbstractTest;
-use PDepend\Source\AST\ASTConstantDeclarator;
-use PDepend\Source\AST\ASTExpression;
-use PDepend\Source\AST\ASTNamespace;
-use PDepend\Source\Builder\Builder;
-use PDepend\Source\Tokenizer\Tokenizer;
-use PDepend\Util\Cache\CacheDriver;
-
-/**
- * Test case for the {@link \PDepend\Source\Language\PHP\PHPParserVersion70} class.
- *
- * @copyright 2008-2017 Manuel Pichler. All rights reserved.
- * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- * @since 2.3
- *
- * @covers \PDepend\Source\Language\PHP\PHPParserVersion70
- * @group unittest
- */
-class PHPParserVersion70Test extends AbstractTest
-{
-    /**
-     * testFormalParameterScalarTypeHintInt
-     *
-     * @return void
-     */
-    public function testFormalParameterScalarTypeHintInt()
-    {
-        $type = $this->getFirstFormalParameterForTestCase()->getType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertEquals('int', $type->getImage());
-    }
-
-    /**
-     * testFormalParameterScalarTypeHintString
-     *
-     * @return void
-     */
-    public function testFormalParameterScalarTypeHintString()
-    {
-        $type = $this->getFirstFormalParameterForTestCase()->getType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertEquals('string', $type->getImage());
-    }
-
-    /**
-     * testFormalParameterScalarTypeHintFloat
-     *
-     * @return void
-     */
-    public function testFormalParameterScalarTypeHintFloat()
-    {
-        $type = $this->getFirstFormalParameterForTestCase()->getType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertEquals('float', $type->getImage());
-    }
-
-    /**
-     * testFormalParameterScalarTypeHintBool
-     *
-     * @return void
-     */
-    public function testFormalParameterScalarTypeHintBool()
-    {
-        $type = $this->getFirstFormalParameterForTestCase()->getType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertEquals('bool', $type->getImage());
-    }
-
-    /**
-     * testFormalParameterStillWorksWithTypeHintArray
-     *
-     * @return void
-     */
-    public function testFormalParameterStillWorksWithTypeHintArray()
-    {
-        $type = $this->getFirstFormalParameterForTestCase()->getChild(0);
-
-        $this->assertFalse($type->isScalar());
-    }
-
-    /**
-     * testFunctionReturnTypeHintInt
-     *
-     * @return void
-     */
-    public function testFunctionReturnTypeHintInt()
-    {
-        $type = $this->getFirstFunctionForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertSame('int', $type->getImage());
-    }
-
-    /**
-     * testFunctionReturnTypeHintFloat
-     *
-     * @return void
-     */
-    public function testFunctionReturnTypeHintFloat()
-    {
-        $type = $this->getFirstFunctionForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertSame('float', $type->getImage());
-    }
-
-    /**
-     * testFunctionReturnTypeHintString
-     *
-     * @return void
-     */
-    public function testFunctionReturnTypeHintString()
-    {
-        $type = $this->getFirstFunctionForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertSame('string', $type->getImage());
-    }
-
-    /**
-     * testFunctionReturnTypeHintBool
-     *
-     * @return void
-     */
-    public function testFunctionReturnTypeHintBool()
-    {
-        $type = $this->getFirstFunctionForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertSame('bool', $type->getImage());
-    }
-
-    /**
-     * testFunctionReturnTypeHintArray
-     *
-     * @return void
-     */
-    public function testFunctionReturnTypeHintArray()
-    {
-        $type = $this->getFirstFunctionForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isArray());
-        $this->assertSame('array', $type->getImage());
-    }
-
-    /**
-     * testFunctionReturnTypeHintCallable
-     *
-     * @return void
-     */
-    public function testFunctionReturnTypeHintCallable()
-    {
-        $type = $this->getFirstFunctionForTestCase()->getReturnType();
-
-        $this->assertFalse($type->isScalar());
-        $this->assertFalse($type->isArray());
-
-        $this->assertSame('callable', $type->getImage());
-    }
-
-    /**
-     * testFunctionReturnTypeHintClass
-     *
-     * @return void
-     */
-    public function testFunctionReturnTypeHintClass()
-    {
-        $type = $this->getFirstFunctionForTestCase()->getReturnType();
-
-        $this->assertFalse($type->isScalar());
-        $this->assertFalse($type->isArray());
-
-        $this->assertSame('\\Iterator', $type->getImage());
-    }
-
-    /**
-     * testClosureReturnTypeHintInt
-     *
-     * @return void
-     */
-    public function testClosureReturnTypeHintInt()
-    {
-        $type = $this->getFirstClosureForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertSame('int', $type->getImage());
-    }
-
-    /**
-     * testClosureReturnTypeHintFloat
-     *
-     * @return void
-     */
-    public function testClosureReturnTypeHintFloat()
-    {
-        $type = $this->getFirstClosureForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertSame('float', $type->getImage());
-    }
-
-    /**
-     * testClosureReturnTypeHintString
-     *
-     * @return void
-     */
-    public function testClosureReturnTypeHintString()
-    {
-        $type = $this->getFirstClosureForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertSame('string', $type->getImage());
-    }
-
-    /**
-     * testClosureReturnTypeHintBool
-     *
-     * @return void
-     */
-    public function testClosureReturnTypeHintBool()
-    {
-        $type = $this->getFirstClosureForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isScalar());
-        $this->assertSame('bool', $type->getImage());
-    }
-
-    /**
-     * testClosureReturnTypeHintArray
-     *
-     * @return void
-     */
-    public function testClosureReturnTypeHintArray()
-    {
-        $type = $this->getFirstClosureForTestCase()->getReturnType();
-
-        $this->assertTrue($type->isArray());
-        $this->assertSame('array', $type->getImage());
-    }
-
-    /**
-     * testClosureReturnTypeHintCallable
-     *
-     * @return void
-     */
-    public function testClosureReturnTypeHintCallable()
-    {
-        $type = $this->getFirstClosureForTestCase()->getReturnType();
-
-        $this->assertFalse($type->isScalar());
-        $this->assertFalse($type->isArray());
-
-        $this->assertSame('callable', $type->getImage());
-    }
-
-    /**
-     * testClosureReturnTypeHintClass
-     *
-     * @return void
-     */
-    public function testClosureReturnTypeHintClass()
-    {
-        $type = $this->getFirstClosureForTestCase()->getReturnType();
-
-        $this->assertFalse($type->isScalar());
-        $this->assertFalse($type->isArray());
-
-        $this->assertSame('\\Iterator', $type->getImage());
-    }
-
-    /**
-     * testSpaceshipOperatorWithStrings
-     *
-     * @return void
-     */
-    public function testSpaceshipOperatorWithStrings()
-    {
-        $expr = $this->getFirstClassMethodForTestCase()
-            ->getFirstChildOfType('PDepend\\Source\\AST\\ASTExpression')
-            ->getFirstChildOfType('PDepend\\Source\\AST\\ASTExpression');
-
-        $this->assertSame('<=>', $expr->getImage());
-    }
-
-    /**
-     * testSpaceshipOperatorWithNumbers
-     *
-     * @return void
-     */
-    public function testSpaceshipOperatorWithNumbers()
-    {
-        $expr = $this->getFirstClassMethodForTestCase()
-            ->getFirstChildOfType('PDepend\\Source\\AST\\ASTExpression')
-            ->getFirstChildOfType('PDepend\\Source\\AST\\ASTExpression');
-
-        $this->assertSame('<=>', $expr->getImage());
-    }
-
-    /**
-     * testSpaceshipOperatorWithArrays
-     *
-     * @return \PDepend\Source\AST\ASTExpression
-     */
-    public function testSpaceshipOperatorWithArrays()
-    {
-        $expr = $this->getFirstClassMethodForTestCase()
-            ->getFirstChildOfType('PDepend\\Source\\AST\\ASTExpression')
-            ->getChild(1);
-
-        $this->assertSame('<=>', $expr->getImage());
-
-        return $expr;
-    }
-
-    /**
-     * @param \PDepend\Source\AST\ASTExpression $expr
-     * @return void
-     * @depends testSpaceshipOperatorWithArrays
-     */
-    public function testSpaceshipOperatorHasExpectedStartLine(ASTExpression $expr)
-    {
-        $this->assertSame(6, $expr->getStartLine());
-    }
-
-    /**
-     * @param \PDepend\Source\AST\ASTExpression $expr
-     * @return void
-     * @depends testSpaceshipOperatorWithArrays
-     */
-    public function testSpaceshipOperatorHasExpectedEndLine(ASTExpression $expr)
-    {
-        $this->assertSame(6, $expr->getEndLine());
-    }
-
-    /**
-     * @param \PDepend\Source\AST\ASTExpression $expr
-     * @return void
-     * @depends testSpaceshipOperatorWithArrays
-     */
-    public function testSpaceshipOperatorHasExpectedStartColumn(ASTExpression $expr)
-    {
-        $this->assertSame(27, $expr->getStartColumn());
-    }
-
-    /**
-     * @param \PDepend\Source\AST\ASTExpression $expr
-     * @return void
-     * @depends testSpaceshipOperatorWithArrays
-     */
-    public function testSpaceshipOperatorHasExpectedEndColumn(ASTExpression $expr)
-    {
-        $this->assertSame(29, $expr->getEndColumn());
-    }
-
-    /**
-     * testNullCoalesceOperator
-     *
-     * @return void
-     */
-    public function testNullCoalesceOperator()
-    {
-        $expr = $this->getFirstClassMethodForTestCase()
-            ->getFirstChildOfType('PDepend\\Source\\AST\\ASTExpression')
-            ->getFirstChildOfType('PDepend\\Source\\AST\\ASTExpression');
-
-        $this->assertSame('??', $expr->getImage());
-    }
-
-    /**
-     * @return void
-     */
-    public function testListKeywordAsMethodName()
-    {
-        $method = $this->getFirstMethodForTestCase();
-        $this->assertNotNull($method);
-    }
-
-    /**
-     * @return void
-     * @expectedException \PDepend\Source\Parser\UnexpectedTokenException
-     */
-    public function testListKeywordAsFunctionNameThrowsException()
-    {
-        $this->parseCodeResourceForTest();
-    }
-
-    /**
-     * @return \PDepend\Source\AST\ASTNamespace
-     */
-    public function testGroupUseStatement()
-    {
-        $namespaces = $this->parseCodeResourceForTest();
-        $this->assertNotNull($namespaces);
-
-        return $namespaces[0];
-    }
-
-    /**
-     * @param \PDepend\Source\AST\ASTNamespace $namespace
-     * @return void
-     * @depends testGroupUseStatement
-     */
-    public function testGroupUseStatementClassNameResolution(ASTNamespace $namespace)
-    {
-        $classes = $namespace->getClasses();
-        $class = $classes[0];
-
-        $this->assertEquals(
-            'FooLibrary\Bar\Baz\ClassB',
-            $class->getParentClass()->getNamespacedName()
-        );
-    }
-
-    /**
-     * @param \PDepend\Source\AST\ASTNamespace $namespace
-     * @return void
-     * @depends testGroupUseStatement
-     */
-    public function testGroupUseStatementAliasResolution(ASTNamespace $namespace)
-    {
-        $classes = $namespace->getClasses();
-        $class = $classes[1];
-
-        $this->assertEquals(
-            'FooLibrary\Bar\Baz\ClassD',
-            $class->getParentClass()->getNamespacedName()
-        );
-    }
-
-    /**
-     * @return void
-     */
-    public function testUniformVariableSyntax()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    /**
-     * @return void
-     */
-    public function testConstantNameArray()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    /**
-     * @return void
-     */
-    public function testClassConstantNames()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    /**
-     * @return void
-     */
-    public function testClassConstantNamesAccessed()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    /**
-     * @return void
-     */
-    public function testClassMethodNames()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    /**
-     * @return void
-     */
-    public function testClassMethodNamesInvoked()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    /**
-     * @return void
-     */
-    public function testYieldFrom()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    /**
-     * @return void
-     */
-    public function testParseList()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    /**
-     * @return void
-     */
-    public function testParseListWithSquaredBrackets()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: [, line: 2, col: 26, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    /**
-     * Tests that the parser throws an exception when it detects an invalid
-     * token in a method or property declaration.
-     *
-     * @return void
-     */
-    public function testParserThrowsUnexpectedTokenExceptionForInvalidTokenInPropertyDeclaration()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: const, line: 4, col: 13, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    /**
-     * Tests that the parser throws an exception when using :void on PHP < 7.1.
-     *
-     * @return void
-     */
-    public function testVoidTypeHintReturn()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: void, line: 2, col: 23, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    /**
-     * Tests that the parser throws an exception when using [...] = ... with PHP < 7.1.
-     *
-     * @return void
-     */
-    public function testListExpressionWithSquaredBrackets()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: [, line: 4, col: 5, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    /**
-     * testListExpressionWithKeys
-     *
-     * @return void
-     */
-    public function testListExpressionWithKeys()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: "a", line: 4, col: 10, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    /**
-     * testListExpressionWithKeysAndNestedList
-     *
-     * @return void
-     */
-    public function testListExpressionWithKeysAndNestedList()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: \'a\', line: 4, col: 10, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    /**
-     * @param \PDepend\Source\Tokenizer\Tokenizer $tokenizer
-     * @param \PDepend\Source\Builder\Builder<mixed> $builder
-     * @param \PDepend\Util\Cache\CacheDriver $cache
-     * @return \PDepend\Source\Language\PHP\AbstractPHPParser
-     */
-    protected function createPHPParser(Tokenizer $tokenizer, Builder $builder, CacheDriver $cache)
-    {
-        return $this->getAbstractClassMock(
-            'PDepend\\Source\\Language\\PHP\\PHPParserVersion70',
-            array($tokenizer, $builder, $cache)
-        );
-    }
-
-    public function testParenthesisAroundCallableParsesArguments()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    public function testKeywordsAsMethodNames()
-    {
-        $namespaces = $this->parseCodeResourceForTest();
-
-        $classes = $namespaces[0]->getClasses();
-        $methods = $classes[0]->getMethods();
-
-        $this->assertSame('trait', $methods[0]->getName());
-        $this->assertSame('callable', $methods[1]->getName());
-        $this->assertSame('insteadof', $methods[2]->getName());
-    }
-
-    public function testKeywordsAsConstants()
-    {
-        $namespaces = $this->parseCodeResourceForTest();
-
-        $classes = $namespaces[0]->getClasses();
-        /** @var ASTConstantDeclarator[] $constants */
-        $constants = $classes[0]->findChildrenOfType('PDepend\\Source\\AST\\ASTConstantDeclarator');
-
-        $this->assertSame('trait', $constants[0]->getImage());
-        $this->assertSame('callable', $constants[1]->getImage());
-        $this->assertSame('insteadof', $constants[2]->getImage());
-    }
-
-    public function testCallableKeywordAsClassName()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: callable, line: 3, col: 7, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    public function testTraitKeywordAsClassName()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: trait, line: 3, col: 7, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    public function testInsteadofKeywordAsClassName()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: insteadof, line: 3, col: 7, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    public function testCallableKeywordAsInterfaceName()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: callable, line: 3, col: 11, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    public function testTraitKeywordAsInterfaceName()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: trait, line: 3, col: 11, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    public function testInsteadofKeywordAsInterfaceName()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\UnexpectedTokenException',
-            'Unexpected token: insteadof, line: 3, col: 11, file: '
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-
-    /**
-     * Tests that the parser does not throw an exception when it detects a reserved
-     * keyword in constant class names.
-     *
-     * @return void
-     */
-    public function testReservedKeyword()
-    {
-        $this->assertNotNull($this->parseCodeResourceForTest());
-    }
-
-    /**
-     * testSymmetricArrayDestructuringEmptySlot
-     *
-     * @return void
-     */
-    public function testSymmetricArrayDestructuringEmptySlot()
-    {
-        $this->setExpectedException(
-            '\\PDepend\\Source\\Parser\\InvalidStateException',
-            'The parser has reached an invalid state near line "4" in file'
-        );
-
-        $this->parseCodeResourceForTest();
-    }
-}
